feat(chrome-extension): move dropdown to separate feature
