+++ apps/chrome-extension/src/features/textAdaptationPropfilesDropdown/index.ts
@@ -1,0 +1,1 @@
+export { default as TextAdaptationProfilesDropdown } from './ui/TextAdaptationProfilesDropdown.vue'
